transport: remove Open() from the transport interface
This abstracts out serial port/tcp socket management from transports,
allowing them to be used on any object implemeting a subset of the
net.Conn interface.

The end-goal is to allow the tcp transport to be used on both
client or server sockets. Added benefit: having the connection
management logic in the client (and later on in the server) rather
than in transports makes for clearer code and interfaces.